Created Progressbar and Piechart component with ProgressBar.scss
