+++ package.json
@@ -15,11 +15,13 @@
     "@types/react-router-dom": "^5.3.3",
     "antd": "^5.20.3",
     "chart.js": "^4.4.4",
+    "date-fns": "^3.6.0",
     "lodash": "^4.17.21",
     "react": "^18.3.1",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.3.1",
     "react-redux": "^9.1.2",
+    "react-router-dom": "^6.26.1",
     "recharts": "^2.12.7",
     "sass": "^1.77.8"
   },
