+++ src/components/HeaderComponent/Header.module.css
@@ -1,0 +1,64 @@
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.nav {
+  display: flex;
+  gap: 20px;
+  align-items: center;
+}
+
+.nav-item1 {
+  color: var(--Text, #4c506b);
+  font-feature-settings: "liga" off, "clig" off;
+  font-family: "DM Sans";
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+}
+
+.nav-item2 {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.child-item1 {
+  color: var(--Crypto-purple, #453de0);
+  font-feature-settings: "liga" off, "clig" off;
+  font-family: "DM Sans";
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+}
+
+.child-item2 {
+  color: var(--Label, #a9acbb);
+  font-feature-settings: "liga" off, "clig" off;
+  font-family: "DM Sans";
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+}
+
+.vertical-line {
+  width: 1px;
+  height: 37px;
+  background: #a9acbb;
+  margin: 0 10px;
+}
+
+.data {
+  background-color: white;
+}
+
+.wallet-connect {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}/*# sourceMappingURL=Header.module.css.map */
