+++ src/components/Maincontent/ProgressBar.scss
@@ -1,0 +1,68 @@
+.progress-card {
+    width: 18.75rem; // 300px
+  }
+  
+  .progress-info {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+  }
+  
+  .progress-title {
+    color: #9FA2B2;
+    font-feature-settings: 'liga' off, 'clig' off;
+    font-family: 'DM Sans';
+    font-size: 1.125rem; // 18px
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+  }
+  
+  .expand-limit-button {
+    color: #453DE0;
+    font-feature-settings: 'liga' off, 'clig' off;
+    font-family: 'DM Sans';
+    font-size: 1.209rem; // 19.355px
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+    padding: 0;
+  }
+  
+  .progress-circle {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  
+  .circle-container {
+    width: 7.5625rem; // 121px
+    height: 7.5625rem; // 121px
+    position: relative;
+  }
+  
+  .progress-format {
+    font-size: 0.75rem; // 12px
+    text-align: center;
+  }
+  
+  .progress-percent {
+    color: #4C506B;
+    font-feature-settings: 'liga' off, 'clig' off;
+    font-family: 'DM Sans';
+    font-size: 1.548rem; // 24.763px
+    font-style: normal;
+    font-weight: 700;
+    line-height: 1.45; // 23.574px / 16px
+  }
+  
+  .progress-label {
+    color: #9FA2B2;
+    font-feature-settings: 'liga' off, 'clig' off;
+    font-family: 'DM Sans';
+    font-size: 0.844rem; // 13.507px
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+  }
+
