+++ src/components/Maincontent/mainPage.tsx
@@ -1,29 +1,37 @@
-import React, { useState, useEffect } from 'react';
-import { LineChart, Line, ResponsiveContainer } from 'recharts';
-import { Row, Col, Card, Tabs } from 'antd';
-import ProgressBar from './ProgressBar';
+import React, { useState, useEffect } from "react";
+import { LineChart, Line, ResponsiveContainer } from "recharts";
+import { Row, Col, Card, Tabs } from "antd";
+import PieChartCircle from "./PieChartCircle";
+import ProgressBar from "./ProgressBar";
+
+
 
 const { TabPane } = Tabs;
 
 const MainPage: React.FC = () => {
-  const [activeKey, setActiveKey] = useState('today');
+  const [activeKey, setActiveKey] = useState("today");
   const [data, setData] = useState<any>({});
 
   // Fetch data from db.json
   useEffect(() => {
-    fetch('/db.json')
+    fetch("/db.json")
       .then((response) => response.json())
       .then((data) => setData(data.data)) // Adjust if your JSON structure is different
-      .catch((error) => console.error('Error fetching data:', error));
+      .catch((error) => console.error("Error fetching data:", error));
   }, []);
 
   return (
-    <Row gutter={16} style={{ marginTop: 20, padding: 30 }}>
+    <Row gutter={18} style={{ marginTop: 10, padding: 10 }}>
       <Col span={10}>
         <Card title="Your Balance">
           <ResponsiveContainer width="100%" height={150}>
             <LineChart data={data[activeKey]}>
-              <Line type="monotone" dataKey="pv" stroke="#8884d8" strokeWidth={2} />
+              <Line
+                type="monotone"
+                dataKey="pv"
+                stroke="#8884d8"
+                strokeWidth={1}
+              />
             </LineChart>
           </ResponsiveContainer>
           <Tabs defaultActiveKey="today" onChange={(key) => setActiveKey(key)}>
@@ -35,12 +43,13 @@ const MainPage: React.FC = () => {
         </Card>
       </Col>
       <Col span={8}>
-          <ProgressBar />
+        <PieChartCircle />
+      </Col>
+      <Col span={4}>
+        <ProgressBar/>
       </Col>
     </Row>
-    
   );
-
 };
 
 export default MainPage;
